add docs and better structure
