+++ load-testing/scenarios/requestBlockNumberEtherium.js
@@ -1,22 +1,20 @@
 // Import necessary modules
-import { sleep, check } from 'k6';
+import { check } from 'k6';
 import { sendEthereumRequest } from '../modules/etheriumRequests.js';
-import { ENV_CONFIG } from '../config/env.js';
 
-// Anvil direct test scenario
-export function anvilDirectScenario() {
+export function requestBlockNumberEtheriumScenario(baseUrl) {
     // Example method and parameters for the Ethereum JSON-RPC request
     const method = "eth_blockNumber";
     const params = [];
-    const baseUrl = ENV_CONFIG.anvilBaseUrl;
 
-    const tags = {
-        method: method,
-        baseUrl: baseUrl,
-    }
+    // We can populate tags in addition to the defaults.
+    // const tags = {
+    //     method: method,
+    //     baseUrl: baseUrl,
+    // }
 
     // Send request and receive response
-    let response = sendEthereumRequest(baseUrl, method, params, tags);
+    let response = sendEthereumRequest(baseUrl, method, params);
 
     // Additional checks specific to the scenario can be written like that:
     // check(response, {
@@ -25,7 +23,4 @@ export function anvilDirectScenario() {
     //         return jsonResponse.hasOwnProperty("result") && jsonResponse.hasOwnProperty("id");
     //     },
     // }, tags);
-
-    // Simulate think time
-    sleep(1);
 }
