+++ load-testing/tests/anvilDirectEtherium.js
@@ -1,0 +1,17 @@
+// Our custom handleSummary produces HTML report with CLI output and some default configuration.
+export { handleSummary, options } from '../config/index.js';
+import { requestBlockNumberEtheriumScenario } from '../scenarios/requestBlockNumberEtherium.js';
+import { ENV_CONFIG } from '../config/index.js';
+import { sleep } from 'k6';
+
+
+// The function that defines VU logic.
+//
+// See https://grafana.com/docs/k6/latest/examples/get-started-with-k6/ to learn more
+// about authoring k6 scripts.
+export default function () {
+    requestBlockNumberEtheriumScenario(ENV_CONFIG.anvilBaseUrl);
+
+    // Simulate think time
+    sleep(1);
+}
